move components to individual folders
