+++ client/src/components/miniHistogram/index.js
