+++ client/src/components/miniStackedBar/index.js
